+++ rakaia/plugins/__init__.py
@@ -1,7 +1,9 @@
 from typing import Union
 from functools import partial
 import plotly.graph_objs as go
-from rakaia.plugins.models import leiden_clustering
+from rakaia.plugins.models import (
+    QuantificationRandomForest,
+    leiden_clustering)
 
 class PluginNotFoundError(Exception):
     """
@@ -15,7 +17,7 @@ class PluginDescriptors:
     :return: None
     """
     # the descriptors should be the string representation of the model in the application dropdown
-    descriptors = ["leiden"]
+    descriptors = ["leiden", "random forest"]
 
 class PluginModes:
     """
@@ -30,7 +32,29 @@ class PluginModes:
     @staticmethod
     def leiden(quantification_results: Union[dict, go.Figure], input_category: str,
                   output_category: str):
+        """
+        Generate leiden clustering
+
+        :param quantification_results: `pd.DataFrame` or `anndata.AnnData` object containing tabular object intensity measurements
+        :param input_category: Placeholder argument for similar plugins. NOt used for leiden clustering
+        :param output_category: Column to store the outputs of the clustering procedure
+        :return: Quantification results with the cluster labels added under the output column.
+        """
         return leiden_clustering(quantification_results, output_category)
+
+    @staticmethod
+    def random_forest(quantification_results: Union[dict, go.Figure], input_category: str,
+                  output_category: str):
+        """
+        Train and predict using a rando forest classifier
+
+        :param quantification_results: `pd.DataFrame` or `anndata.AnnData` object containing tabular object intensity measurements
+        :param input_category: Existing annotation column in the object that specifies the classes used to fit the model
+        :param output_category: Column to store the output labels of the classifier prediction
+        :return: Quantification results with the prediction labels added under the output column.
+        """
+        return QuantificationRandomForest(
+            quantification_results, input_category, output_category).quantification_with_labels(True)
 
 
 class PluginModelModes:
@@ -41,6 +65,7 @@ class PluginModelModes:
     """
     # each attribute should be directly named after a specific function in PluginModes, and should be a partial function
     leiden = partial(PluginModes.leiden)
+    random_forest = partial(PluginModes.random_forest)
 
 def run_quantification_model(quantification_results: Union[dict, go.Figure], input_category: Union[str, None]=None,
                              output_category: str="out", mode: str="leiden"):
@@ -49,9 +74,9 @@ def run_quantification_model(quantification_results: Union[dict, go.Figure], inp
     mode provided by the user
     """
     # convert the mode to a spaceless string representation to match the attribute
-    mode = mode.replace(" ", "_") if mode else ""
     if mode not in PluginDescriptors.descriptors:
         raise PluginNotFoundError(f"The plugin mode provided: {mode} is not a supported plugin."
                                   f"Current plugins include: {PluginDescriptors.descriptors}")
+    mode = mode.replace(" ", "_") if mode else ""
     with_model = getattr(PluginModelModes, mode)(quantification_results, input_category, output_category)
     return with_model