+++ tests/test_plugins.py
@@ -1,11 +1,14 @@
 import os
 import anndata
 import pandas as pd
+import numpy as np
 import pytest
 from rakaia.plugins import (
     PluginNotFoundError,
     run_quantification_model)
-from rakaia.plugins.models import leiden_clustering
+from rakaia.plugins.models import (
+    QuantificationRandomForest,
+    leiden_clustering)
 
 
 def test_leiden_clustering(get_current_dir):
@@ -17,10 +20,20 @@ def test_leiden_clustering(get_current_dir):
     assert isinstance(with_leiden_adata, anndata.AnnData)
     assert "leiden" in with_leiden_adata.obs.columns
 
+def test_quant_random_forest(get_current_dir):
+    measurements_csv = pd.read_csv(os.path.join(get_current_dir, "cell_measurements.csv"))
+    # predict the area for test_2
+    measurements_csv['area_use'] = np.where(measurements_csv['sample'] == 'test_1', measurements_csv['area'], 'Unassigned')
+    with_predictions = QuantificationRandomForest(measurements_csv, "area_use", "area_predict").quantification_with_labels()
+    assert 'area_predict' in pd.DataFrame(with_predictions).columns
+    assert 'Unassigned' not in pd.DataFrame(with_predictions)['area_predict']
+
 def test_run_quant_models(get_current_dir):
     measurements_csv = pd.read_csv(os.path.join(get_current_dir, "cell_measurements.csv"))
     leiden_as_model = run_quantification_model(measurements_csv, None)
     assert "out" in pd.DataFrame(leiden_as_model).columns
+    rf_model = run_quantification_model(leiden_as_model, "area", "area_predict", "random forest")
+    assert "area_predict" in pd.DataFrame(rf_model).columns
 
     with pytest.raises(PluginNotFoundError):
         run_quantification_model(measurements_csv, None, mode="not found")